+++ src/stock/websocket.py
@@ -178,92 +178,6 @@ async def run_websocket_background_multiple(stock_symbols: List[Dict[str, str]])
 
 
 
-event_loop = asyncio.get_event_loop()
-
-# SSE 전용 WebSocket 스레드 실행 함수
-def sse_websocket_thread(stock_symbols, data_queue, loop):
-    logger.info("Starting SSE WebSocket thread for symbols: %s", stock_symbols)
-
-    def on_open_wrapper(ws):
-        on_open(ws, stock_symbols)
-
-    def handle_sse_message(ws, message):
-        if message.startswith("{"):
-            return
-        else:
-            d1 = message.split("|")
-            if len(d1) >= 3:
-                stock_symbol = d1[3].split("^")[0]
-                stock_data = process_data_for_sse(message, stock_symbol)  # SSE용 데이터 처리
-                if stock_data:
-                    # 메인 스레드의 이벤트 루프를 사용하여 Queue에 추가
-                    asyncio.run_coroutine_threadsafe(data_queue.put(json.dumps(stock_data)), loop)
-                    print(f"Added data for symbol {stock_symbol} to data_queue.")
-
-    while True:
-        try:
-            ws = websocket.WebSocketApp(
-                "ws://ops.koreainvestment.com:31000",
-                on_open=on_open_wrapper,
-                on_message=lambda ws, message: handle_sse_message(ws, message),
-                on_error=on_error,
-                on_close=on_close
-            )
-            ws.run_forever(ping_interval=30)
-            logger.info("SSE WebSocket thread has been terminated.")
-        except Exception as e:
-            logger.error(f"SSE WebSocket error occurred: {e}")
-            time.sleep(0.3)
-            logger.info("Attempting to reconnect SSE WebSocket...")
-
-# WebSocket 백그라운드 실행 함수 (SSE 전용)
-async def run_websocket_background_to_queue(stock_symbols: List[Dict[str, str]]) -> asyncio.Queue:
-    data_queue = asyncio.Queue()
-    # 메인 이벤트 루프 참조 전달
-    ws_thread = threading.Thread(target=sse_websocket_thread, args=(stock_symbols, data_queue, event_loop))
-    ws_thread.start()
-    return data_queue
-
-
-# SSE 전송을 위한 데이터 처리 함수
-def process_data_for_sse(data, stock_symbol):
-    stock_info = get_company_details(stock_symbol)
-    if not stock_info or "id" not in stock_info or "name" not in stock_info:
-        print(f"No valid company information for symbol: {stock_symbol}")
-        return None
-
-    try:
-        d1 = data.split("|")
-        if len(d1) >= 4:
-            recvData = d1[3]
-            result = recvData.split("^")
-            if len(result) > 12:
-                stock_data = {
-                    "id": stock_info.get("id"),
-                    "name": stock_info.get("name"),
-                    "symbol": stock_symbol,
-                    "close": result[2],
-                    "rate_price": result[4],
-                    "rate": result[5],
-                    "volume": result[12],
-                }
-                return stock_data
-            else:
-                print(f"Unexpected result format for data: {result}")
-    except (IndexError, ValueError, TypeError) as e:
-        print(f"Error processing stock data for SSE: {e}")
-    return None
-
-
-
-
-
-
-
-
-
-
-
 # Mock 데이터 생성 함수 - 개별 주식 데이터 생성
 def generate_single_mock_stock_data(stock_info: Dict[str, str]) -> Dict[str, str]:
     # 주식 정보 조회 및 목업 데이터 생성
